Rename aux to regex and update fleetapi token refresh error checking
